requirements.txt
streamlit
beautifulsoup4
requests
crewai
python-dotenv
langchain-anthropic
pandas
nltk